Added lua scripts importing

+++ vn_engine/ResizeEvent.cpp
@@ -1,6 +1,6 @@
 #include "ResizeEvent.h"
 
-ResizeEvent::ResizeEvent(int old_x, int old_y, int new_x, int new_y, int _scaleX, int _scaleY)
+ResizeEvent::ResizeEvent(int old_x, int old_y, int new_x, int new_y, double _scaleX, double _scaleY)
 {
 	old_width = old_x;
 	old_height = old_y;
